Make the UI an actor
