+++ plugins/tui/include/tui/ui.hpp
@@ -8,38 +8,20 @@
 
 #pragma once
 
-#include "tui/tui.hpp"
+#include "vast/atoms.hpp"
+#include <memory>
 
-#include <vast/atoms.hpp>
-
-#include <caf/typed_event_based_actor.hpp>
-
-#include <vector>
+#include <caf/typed_actor.hpp>
 
 namespace vast::plugins::tui {
 
 using ui_actor = caf::typed_actor<
-  // Initiates the main UI loop.
+  /// Receive a log message.
   caf::reacts_to<std::string>,
+  /// Kick off the UI main loop.
   caf::reacts_to<atom::run>>;
 
-/// @relates ui
-struct ui_state {
-  ui_state(ui_actor::stateful_pointer<ui_state> self);
-
-  /// Hook ourselves into the system-wide logger.
-  void hook_logger();
-
-  class tui tui;
-
-  /// Points to the parent actor.
-  ui_actor::pointer self;
-
-  /// Gives this actor a recognizable name in logging output.
-  static inline const char* name = "ui";
-};
-
-/// Renders the UI.
-ui_actor::behavior_type ui(ui_actor::stateful_pointer<ui_state> self);
+/// Spawns the UI actor.
+ui_actor spawn_ui(caf::actor_system& system);
 
 } // namespace vast::plugins::tui
